+++ com.jaspersoft.studio.data.sql/src-gen/com/jaspersoft/studio/data/sql/impl/GroupByColumnFullImpl.java
@@ -60,7 +60,7 @@ public class GroupByColumnFullImpl extends OrGroupByColumnImpl implements GroupB
    * @generated
    * @ordered
    */
-  protected static final Integer GR_BY_INT_EDEFAULT = null;
+  protected static final Long GR_BY_INT_EDEFAULT = null;
 
   /**
    * The cached value of the '{@link #getGrByInt() <em>Gr By Int</em>}' attribute.
@@ -70,7 +70,7 @@ public class GroupByColumnFullImpl extends OrGroupByColumnImpl implements GroupB
    * @generated
    * @ordered
    */
-  protected Integer grByInt = GR_BY_INT_EDEFAULT;
+  protected Long grByInt = GR_BY_INT_EDEFAULT;
 
   /**
    * <!-- begin-user-doc -->
@@ -194,7 +194,7 @@ public class GroupByColumnFullImpl extends OrGroupByColumnImpl implements GroupB
    * <!-- end-user-doc -->
    * @generated
    */
-  public Integer getGrByInt()
+  public Long getGrByInt()
   {
     return grByInt;
   }
@@ -204,9 +204,9 @@ public class GroupByColumnFullImpl extends OrGroupByColumnImpl implements GroupB
    * <!-- end-user-doc -->
    * @generated
    */
-  public void setGrByInt(Integer newGrByInt)
+  public void setGrByInt(Long newGrByInt)
   {
-    Integer oldGrByInt = grByInt;
+    Long oldGrByInt = grByInt;
     grByInt = newGrByInt;
     if (eNotificationRequired())
       eNotify(new ENotificationImpl(this, Notification.SET, SqlPackage.GROUP_BY_COLUMN_FULL__GR_BY_INT, oldGrByInt, grByInt));
@@ -267,7 +267,7 @@ public class GroupByColumnFullImpl extends OrGroupByColumnImpl implements GroupB
         setGbFunction((OpFunction)newValue);
         return;
       case SqlPackage.GROUP_BY_COLUMN_FULL__GR_BY_INT:
-        setGrByInt((Integer)newValue);
+        setGrByInt((Long)newValue);
         return;
     }
     super.eSet(featureID, newValue);
