+++ com.jaspersoft.studio.data.sql/src-gen/com/jaspersoft/studio/data/sql/impl/OrderByColumnFullImpl.java
@@ -49,7 +49,7 @@ public class OrderByColumnFullImpl extends OrOrderByColumnImpl implements OrderB
    * @generated
    * @ordered
    */
-  protected static final Integer COL_ORDER_INT_EDEFAULT = null;
+  protected static final Long COL_ORDER_INT_EDEFAULT = null;
 
   /**
    * The cached value of the '{@link #getColOrderInt() <em>Col Order Int</em>}' attribute.
@@ -59,7 +59,7 @@ public class OrderByColumnFullImpl extends OrOrderByColumnImpl implements OrderB
    * @generated
    * @ordered
    */
-  protected Integer colOrderInt = COL_ORDER_INT_EDEFAULT;
+  protected Long colOrderInt = COL_ORDER_INT_EDEFAULT;
 
   /**
    * The default value of the '{@link #getDirection() <em>Direction</em>}' attribute.
@@ -155,7 +155,7 @@ public class OrderByColumnFullImpl extends OrOrderByColumnImpl implements OrderB
    * <!-- end-user-doc -->
    * @generated
    */
-  public Integer getColOrderInt()
+  public Long getColOrderInt()
   {
     return colOrderInt;
   }
@@ -165,9 +165,9 @@ public class OrderByColumnFullImpl extends OrOrderByColumnImpl implements OrderB
    * <!-- end-user-doc -->
    * @generated
    */
-  public void setColOrderInt(Integer newColOrderInt)
+  public void setColOrderInt(Long newColOrderInt)
   {
-    Integer oldColOrderInt = colOrderInt;
+    Long oldColOrderInt = colOrderInt;
     colOrderInt = newColOrderInt;
     if (eNotificationRequired())
       eNotify(new ENotificationImpl(this, Notification.SET, SqlPackage.ORDER_BY_COLUMN_FULL__COL_ORDER_INT, oldColOrderInt, colOrderInt));
@@ -246,7 +246,7 @@ public class OrderByColumnFullImpl extends OrOrderByColumnImpl implements OrderB
         setColOrder((ColumnFull)newValue);
         return;
       case SqlPackage.ORDER_BY_COLUMN_FULL__COL_ORDER_INT:
-        setColOrderInt((Integer)newValue);
+        setColOrderInt((Long)newValue);
         return;
       case SqlPackage.ORDER_BY_COLUMN_FULL__DIRECTION:
         setDirection((String)newValue);
